tooling: widen ruff to the rules that find defects, and fix what it found
select is now F, E7, E9, B, SIM, UP, RUF, BLE - every rule that finds a
defect or a real simplification. 111 findings, all fixed; `ruff check .`
is clean.

Three rules are dropped as noise on this codebase rather than papered
over with per-line noqa, each with the reason written into the config:
E501 (enforcing it now would reformat files wholesale, which is exactly
what this config is not for), E402 (structural - gi.require_version()
must run before `from gi.repository import ...`, so every GTK module
trips it), and RUF001 (the user-facing strings deliberately use em
dashes and µW). E702 is confined entirely to tests/test_payload.py,
whose fake helper is compact one-line stubs on purpose, so it is a
per-file ignore rather than a global drop.

BLE is the interesting addition. The tree already documented every
intentional broad except with `# noqa: BLE001`, so enabling it makes
those markers mean something and flags the next one that isn't
deliberate. RUF100 then removed 56 directives that had stopped doing
anything, and `external` lists the few codes (N802 on the GObject
vfuncs, S310/S606 on the reviewed urllib/subprocess calls) whose markers
document a decision worth keeping.

Beyond dead directives the fixes were: contextlib.suppress for nine
try/except/pass blocks (with the import placed properly, not where the
autofix wanted it), `open(p, "r")` -> `open(p)`, quoted annotations
unquoted now everything has `from __future__ import annotations`,
`typing.Callable` -> `collections.abc`, ClassVar on two mutable class
attributes, strict= on two zip()s, and a parenthesised precedence
chain in the handheld DMI match that read ambiguously.

CI now runs ruff over the whole tree instead of src/helper/tests, which
is how an unused import had been sitting in scripts/make-demo.py.

Verified after: 281 tests, the GUI smoke test, the helper's interface
XML and constants, and the daemon starting and serving over a real bus.

+++ src/goblinmode/gui/page_diagnostics.py
@@ -10,9 +10,9 @@ import gi
 
 gi.require_version("Gtk", "4.0")
 gi.require_version("Adw", "1")
-from gi.repository import Adw, Gdk, Gtk  # noqa: E402
+from gi.repository import Adw, Gdk, Gtk
 
-from goblinmode.i18n import _  # noqa: E402
+from goblinmode.i18n import _
 
 from goblinmode.gui.widgets.buttonrow import button_row
 from goblinmode.gui.widgets.graph import CorrelationGraph, FpsGraph
@@ -170,7 +170,7 @@ class DiagnosticsPage(Adw.PreferencesPage):
             self._sessions_group.remove(self._sessions_empty)
             self._sessions_empty = None
 
-        mins = int(round((s.get("duration_s") or 0) / 60))
+        mins = round((s.get("duration_s") or 0) / 60)
         bits = [s.get("started", "")[:10], f"{mins} min"]
         if s.get("fps_avg") is not None:
             bits.append(f"avg {s['fps_avg']:.0f} fps")
